requirements.txt
# PaddleOCR implementation requirements

# Core OCR library
paddlepaddle==3.1.0
paddleocr==3.1.0

# Image processing
opencv-python==4.11.0.86
Pillow==10.2.0
numpy==1.26.4

# Data validation
pydantic==2.6.1

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Environment management
python-dotenv==1.0.1